backend/requirements.txt
django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
gunicorn>=21.0
anthropic>=0.39
openai>=1.0
python-dotenv>=1.0
dj-database-url>=2.1
whitenoise>=6.6
python-magic>=0.4.27
